rename BE to config
