+++ .gitignore
@@ -1,4 +1,8 @@
 .DS_Store
+.idea/
+.project
+.pydevproject
+.coverage
 *~
 *.pyc
 *.swp
@@ -9,13 +13,8 @@ web_root/js/compressed.*.js
 web_root/collected_static
 pkgs/newrelic-1\.6\.0\.13
 pkgs/newrelic-1\.6\.0\.13\.tar\.gz
-.idea/
 chromedriver.log
 salad-test
-BE/settings_local.py
-BE/local_untracked.py
-BE/settings/local_untracked.py
-ipyton_input_log_history
 ipython_input_log_history
 .ipython_input_log_history
 *.ipynb
@@ -23,13 +22,14 @@ ipython_input_log_history
 web_root/csvs/*
 web_root/uploads/*
 web_root/images/sites/*
-BE/settings/local_untracked.py
+config/settings_local.py
+config/local_untracked.py
+config/settings/local_untracked.py
 node_modules/*
 web_root/CACHE
 collected_static
 celerybeat-schedule*
 web_root/images/sites/*
-.ipython_input_log_history
 test.sqlite
 media_root/*
 media/*
@@ -38,14 +38,11 @@ bpd/data/epa*
 bpd/data/BEDES*
 **/*-old
 lin/migrations/disabled
-.coverage
 htmlcov/
 src/
 *node_modules*
-/seed/static/vendors/bower_components/fine-uploader/
+/seed/static/vendors/bower_components
 docs/build/
-.project
-.pydevproject
 dump.rdb
 log
 *.log
