+++ config/management/commands/grab_manifest.py
