+++ config/management/commands/sync_static.py
