+++ config/wsgi.py
@@ -2,7 +2,7 @@
 :copyright: (c) 2014 Building Energy Inc
 """
 """
-WSGI config for BE project.
+WSGI config for config project.
 
 This module contains the WSGI application used by Django's development server
 and any production WSGI deployments. It should expose a module-level variable
@@ -18,7 +18,7 @@ framework.
 """
 import os
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "BE.settings.main")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.main")
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
