+++ seed/static/seed/tests/data_upload_modal_ctrl.spec.js
@@ -22,7 +22,7 @@ describe("controller: data_upload_modal_ctrl", function(){
     ];
 
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
