+++ seed/static/seed/tests/dataset_detail_controller.spec.js
@@ -9,7 +9,7 @@ describe("controller: dataset_detail_controller", function(){
     
 
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
