+++ seed/static/seed/tests/dataset_list_controller.spec.js
@@ -10,7 +10,7 @@ describe("controller: dataset_list_controller", function(){
     
 
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
