+++ seed/static/seed/tests/edit_label_modal_controller.spec.js
@@ -20,7 +20,7 @@ describe("controller: edit_label_modal_ctrl", function(){
     ];
 
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
