+++ seed/static/seed/tests/export_modal_controller.spec.js
@@ -6,7 +6,7 @@ describe("controller: export_modal_controller", function(){
     var mock_export_service, scope, controller, modal_state;
     var ctrl, ctrl_scope, modalInstance, timeout;
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
