+++ seed/static/seed/tests/mapping_controller.spec.js
@@ -10,7 +10,7 @@ describe("controller: mapping_controller", function(){
 
 
     // make the seed app available for each test
-    // 'BE.seed' is created in TestFilters.html
+    // 'config.seed' is created in TestFilters.html
     beforeEach(function() {
         module('BE.seed');
     });
