+++ seed/templates/seed/_js_urls.html
@@ -1,6 +1,6 @@
 {% load app_urls %}
 <script>
-        // BE urls
+        // config urls
         window.BE = window.BE || {};
         // new urls
         window.BE.app_urls = {% namespaced_urls %};
